backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt>=4.7.2
requests>=2.26.0
drf-yasg